YaGL: Fixed shader preprocessing
Things like this:

can be present in shader code and the code inside
extension is available on host.

As a workaround we find all GL_xxx_yyy tokens and if it's
not a OpenGL ES extension we replace it with ND_xxx_yyy, thus
making sure that the code inside such #ifdef's won't be compiled.
When GL_xxx_yyy is not a preprocessor token it's not a
problem either, it'll be replaced in all places, thus,
shader will operate correctly

Change-Id: Id23f790c1c2a67f1a188e0161d20832e589a24d0