YaGL: Configurable behavior for glBlitFramebuffer
Currently minifying and magnifying blits in
glBlitFramebuffer are disabled because of host
OpenGL driver problems on many GPUs, they simply crash.
But we want to be able to enable this for
some apps that require it, thus we introduce
YAGL_MIN_MAG_BLITS environment variable.
We currently use this to run nintendo wii games
with dolphin-emu, since dolphin-emu uses minifying
glBlitFramebuffer every frame. For GameCube games however
dophin-emu doesn't use it, so they can be run as usual

Change-Id: Ie11a852bc53700231ce012938a8a531bb4e88394
Signed-off-by: Stanislav Vorobiov <s.vorobiov@samsung.com>